# docs/api/unified-api-spec.yaml
openapi: 3.0.3
info:
  title: SPARC Platform Unified API
  description: |
    Comprehensive API specification for the SPARC (Security Platform with Advanced Real-time Control) system.
    This unified specification covers all 24 microservices providing access control, video management, 
    environmental monitoring, visitor management, mobile credentials, and compliance features.
    
    ## Key Features
    - Multi-tenant architecture with tenant isolation
    - Real-time event processing and notifications
    - Offline resilience with 72-hour operation capability
    - Scalable to 10,000 doors and 1,000 concurrent video streams
    - SOX, HIPAA, and PCI-DSS compliance
    - Mobile credential management with mesh networking
    - Advanced analytics and reporting
    
    ## Authentication
    All endpoints require JWT authentication with tenant-specific scoping.
    Include the `Authorization: Bearer <token>` header and `X-Tenant-ID` header in all requests.
    
    ## Multi-Tenant Usage
    All API endpoints are tenant-scoped. Include the `X-Tenant-ID` header to specify the tenant context.
    Data isolation is enforced at the database and application levels.
  version: 1.0.0
  contact:
    name: SPARC Platform API Support
    email: api-support@sparc-platform.com
  license:
    name: Proprietary
    url: https://sparc-platform.com/license
servers:
  - url: https://api.sparc-platform.com/v1
    description: Production API Server
...
    description: Staging API Server
...
security:
  - BearerAuth: []
  - TenantHeader: []
paths:
...
  /integrations:
    get:
      tags:
        - Integration Service
      summary: List integrations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [ldap, active_directory, hvac, fire_safety, elevator, visitor_system, calendar]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error, syncing]
      responses:
        '200':
          description: Integrations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationListResponse'

    post:
      tags:
        - Integration Service
      summary: Create new integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegrationRequest'
      responses:
        '201':
          description: Integration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Integration'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant identifier for multi-tenant isolation
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
  schemas:
...
    Integration:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [ldap, active_directory, hvac, fire_safety, elevator, visitor_system, calendar]
        status:
          type: string
          enum: [active, inactive, error, syncing]
        configuration:
          type: object
        lastSync:
          type: string
          format: date-time
        syncStatus:
          type: string
          enum: [success, failed, partial, in_progress]
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateIntegrationRequest:
      type: object
      required:
        - name
        - type
        - configuration
      properties:
        name:
          type: string
        type:
          type: string
          enum: [ldap, active_directory, hvac, fire_safety, elevator, visitor_system, calendar]
        configuration:
          type: object
        autoSync:
          type: boolean
          default: true
        syncInterval:
          type: integer
          description: Sync interval in minutes
...
    IntegrationListResponse:
      type: object
      properties:
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/Integration'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...